feat(highlightjs): separate CSS and JS from code
